# Conflicts:
#	app/src/androidTest/java/com/tosslab/jandi/app/ui/account/presenter/AccountHomePresenterImplTest.java
#	app/src/main/java/com/tosslab/jandi/app/network/client/EntityClientManager.java
#	app/src/main/java/com/tosslab/jandi/app/network/client/privatetopic/GroupApiV2Client.java
#	app/src/main/java/com/tosslab/jandi/app/network/client/privatetopic/IGroupApiAuth.java
#	app/src/main/java/com/tosslab/jandi/app/network/client/privatetopic/IGroupApiLoader.java
#	app/src/main/java/com/tosslab/jandi/app/network/client/publictopic/ChannelApiV2Client.java
#	app/src/main/java/com/tosslab/jandi/app/network/client/publictopic/IChannelApiAuth.java
#	app/src/main/java/com/tosslab/jandi/app/network/client/publictopic/IChannelApiLoader.java
#	app/src/main/java/com/tosslab/jandi/app/network/manager/RequestApiManager.java
#	app/src/main/java/com/tosslab/jandi/app/network/manager/apiloader/RestApiLoader.java
#	app/src/main/java/com/tosslab/jandi/app/network/manager/restapiclient/JacksonConvertedAuthRestApiClient.java
#	app/src/main/java/com/tosslab/jandi/app/network/models/ResRoomInfo.java
#	app/src/main/java/com/tosslab/jandi/app/ui/filedetail/views/FileUnshareActivity.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/chat/presenter/MainChatListPresenter.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/chat/presenter/MainChatListPresenterImpl.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/topic/adapter/ExpandableTopicAdapter.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/topic/domain/TopicFolderData.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/topic/domain/TopicItemData.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/topic/model/MainTopicModel.java
#	app/src/main/java/com/tosslab/jandi/app/ui/message/detail/model/TopicDetailModel.java
#	app/src/main/java/com/tosslab/jandi/app/utils/parse/ParseUpdateUtil.java

+++ app/src/androidTest/java/com/tosslab/jandi/app/ui/members/presenter/MembersListPresenterImplTest.java
@@ -24,7 +24,7 @@ public class MembersListPresenterImplTest {
 
     private MembersListPresenterImpl presenter;
     private MembersListPresenterImpl.View mockView;
-    private int topicId;
+    private long topicId;
 
     @Before
     public void setUp() throws Exception {
@@ -45,7 +45,7 @@ public class MembersListPresenterImplTest {
 
     @Test
     public void testOnKickUser() throws Exception {
-        int entityId = BaseInitUtil.getUserIdByEmail(BaseInitUtil.TEST2_EMAIL);
+        long entityId = BaseInitUtil.getUserIdByEmail(BaseInitUtil.TEST2_EMAIL);
         {
             final boolean[] finish = {false};
             Mockito.doAnswer(invocationOnMock -> {
