+++ app/src/main/java/com/tosslab/jandi/app/events/entities/MainSelectTopicEvent.java
@@ -4,14 +4,14 @@ package com.tosslab.jandi.app.events.entities;
  * Created by Steve SeongUg Jung on 15. 7. 15..
  */
 public class MainSelectTopicEvent {
-    private final int selectedEntity;
+    private final long selectedEntity;
 
-    public MainSelectTopicEvent(int selectedEntity) {
+    public MainSelectTopicEvent(long selectedEntity) {
 
         this.selectedEntity = selectedEntity;
     }
 
-    public int getSelectedEntity() {
+    public long getSelectedEntity() {
         return selectedEntity;
     }
 }
