+++ app/src/main/java/com/tosslab/jandi/app/local/orm/repositories/TopicFolderRepository.java
@@ -45,7 +45,7 @@ public class TopicFolderRepository {
 
     public List<ResFolder> getFolders() {
         try {
-            int selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
+            long selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
             Dao<ResFolder, ?> FolderDao = helper.getDao(ResFolder.class);
             return FolderDao.queryBuilder()
                     .orderBy("seq", true)
@@ -61,7 +61,7 @@ public class TopicFolderRepository {
 
     public List<ResFolderItem> getFolderItems() {
         try {
-            int selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
+            long selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
             Dao<ResFolderItem, ?> FolderItemDao = helper.getDao(ResFolderItem.class);
             return FolderItemDao.queryBuilder()
                     .where()
@@ -160,11 +160,11 @@ public class TopicFolderRepository {
         }
     }
 
-    public boolean removeFolder(int folderId) {
+    public boolean removeFolder(long folderId) {
         lock.lock();
         boolean result = false;
         try {
-            int selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
+            long selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
             Dao<ResFolder, ?> resFolderDao = helper.getDao(ResFolder.class);
             DeleteBuilder<ResFolder, ?> deleteBuilder = resFolderDao.deleteBuilder();
             deleteBuilder
@@ -182,7 +182,7 @@ public class TopicFolderRepository {
         }
     }
 
-    public boolean removeFolderItem(int roomId) {
+    public boolean removeFolderItem(long roomId) {
         lock.lock();
         boolean result = false;
         try {
@@ -206,7 +206,7 @@ public class TopicFolderRepository {
         lock.lock();
         boolean result = false;
         try {
-            int selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
+            long selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
             Dao<ResFolder, ?> resFolderDao = helper.getDao(ResFolder.class);
             DeleteBuilder<ResFolder, ?> deleteBuilder = resFolderDao.deleteBuilder();
             deleteBuilder
@@ -227,7 +227,7 @@ public class TopicFolderRepository {
         lock.lock();
         boolean result = false;
         try {
-            int selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
+            long selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
             Dao<ResFolderItem, ?> resFolderItemDao = helper.getDao(ResFolderItem.class);
             DeleteBuilder<ResFolderItem, ?> deleteBuilder = resFolderItemDao.deleteBuilder();
             deleteBuilder
@@ -247,7 +247,7 @@ public class TopicFolderRepository {
     public List<FolderExpand> getFolderExpands() {
         lock.lock();
         try {
-            int selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
+            long selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
             Dao<FolderExpand, ?> dao = helper.getDao(FolderExpand.class);
             return dao.queryBuilder()
                     .where()
@@ -262,10 +262,10 @@ public class TopicFolderRepository {
         return new ArrayList<>();
     }
 
-    public boolean upsertFolderExpands(int folderId, boolean expand) {
+    public boolean upsertFolderExpands(long folderId, boolean expand) {
         lock.lock();
         try {
-            int selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
+            long selectedTeamId = AccountRepository.getRepository().getSelectedTeamId();
             Dao<FolderExpand, ?> dao = helper.getDao(FolderExpand.class);
             FolderExpand data = new FolderExpand();
             data.setExpand(expand);
