+++ app/src/main/java/com/tosslab/jandi/app/network/client/EntityClientManager.java
@@ -40,7 +40,7 @@ public class EntityClientManager {
 
     @RootContext
     Context context;
-    private int selectedTeamId;
+    private long selectedTeamId;
 
     @AfterInject
     void initAuthentication() {
@@ -79,45 +79,45 @@ public class EntityClientManager {
         return RequestApiManager.getInstance().createPrivateGroupByGroupApi(reqCreateTopic);
     }
 
-    public ResCommon joinChannel(int id) throws RetrofitError {
+    public ResCommon joinChannel(long id) throws RetrofitError {
         return RequestApiManager.getInstance().joinTopicByChannelApi(id, new ReqDeleteTopic(selectedTeamId));
     }
 
-    public ResCommon leaveChannel(final int id) throws RetrofitError {
+    public ResCommon leaveChannel(final long id) throws RetrofitError {
         return RequestApiManager.getInstance().leaveTopicByChannelApi(id, new ReqDeleteTopic(selectedTeamId));
     }
 
-    public ResCommon leavePrivateGroup(final int id) throws RetrofitError {
+    public ResCommon leavePrivateGroup(final long id) throws RetrofitError {
         return RequestApiManager.getInstance().leaveGroupByGroupApi(id, new ReqTeam(selectedTeamId));
     }
 
-    public ResCommon modifyChannelName(final int id, String name) throws RetrofitError {
+    public ResCommon modifyChannelName(final long id, String name) throws RetrofitError {
         ReqModifyTopicName entityInfo = new ReqModifyTopicName();
         entityInfo.teamId = selectedTeamId;
         entityInfo.name = name;
         return RequestApiManager.getInstance().modifyPublicTopicNameByChannelApi(entityInfo, id);
     }
 
-    public ResCommon modifyPrivateGroupName(final int id, String name) throws RetrofitError {
+    public ResCommon modifyPrivateGroupName(final long id, String name) throws RetrofitError {
         ReqModifyTopicName entityInfo = new ReqModifyTopicName();
         entityInfo.teamId = selectedTeamId;
         entityInfo.name = name;
         return RequestApiManager.getInstance().modifyGroupNameByGroupApi(entityInfo, id);
     }
 
-    public ResCommon deleteChannel(final int id) throws RetrofitError {
+    public ResCommon deleteChannel(final long id) throws RetrofitError {
         return RequestApiManager.getInstance().deleteTopicByChannelApi(id, new ReqDeleteTopic(selectedTeamId));
     }
 
-    public ResCommon deletePrivateGroup(final int id) throws RetrofitError {
+    public ResCommon deletePrivateGroup(final long id) throws RetrofitError {
         return RequestApiManager.getInstance().deleteGroupByGroupApi(selectedTeamId, id);
     }
 
-    public ResCommon inviteChannel(final int id, final List<Integer> invitedUsers) throws RetrofitError {
+    public ResCommon inviteChannel(final long id, final List<Long> invitedUsers) throws RetrofitError {
         return RequestApiManager.getInstance().invitePublicTopicByChannelApi(id, new ReqInviteTopicUsers(invitedUsers, selectedTeamId));
     }
 
-    public ResCommon invitePrivateGroup(final int id, final List<Integer> invitedUsers) throws RetrofitError {
+    public ResCommon invitePrivateGroup(final long id, final List<Long> invitedUsers) throws RetrofitError {
         return RequestApiManager.getInstance().inviteGroupByGroupApi(id, new ReqInviteTopicUsers(invitedUsers, selectedTeamId));
     }
 
@@ -126,11 +126,11 @@ public class EntityClientManager {
      * Entity 즐겨찾기 등록 / 해제
      * **********************************************************
      */
-    public ResCommon enableFavorite(final int entityId) throws RetrofitError {
+    public ResCommon enableFavorite(final long entityId) throws RetrofitError {
         return RequestApiManager.getInstance().enableFavoriteByStarredEntityApi(new ReqTeam(selectedTeamId), entityId);
     }
 
-    public ResCommon disableFavorite(final int entityId) throws RetrofitError {
+    public ResCommon disableFavorite(final long entityId) throws RetrofitError {
         return RequestApiManager.getInstance().disableFavoriteByStarredEntityApi(selectedTeamId, entityId);
     }
 
@@ -139,19 +139,19 @@ public class EntityClientManager {
      * 사용자 프로필
      * **********************************************************
      */
-    public ResLeftSideMenu.User getUserProfile(final int entityId) throws RetrofitError {
+    public ResLeftSideMenu.User getUserProfile(final long entityId) throws RetrofitError {
         return RequestApiManager.getInstance().getMemberProfileByTeamApi(selectedTeamId, entityId);
     }
 
-    public ResLeftSideMenu.User updateUserProfile(final int entityId, final ReqUpdateProfile reqUpdateProfile) throws RetrofitError {
+    public ResLeftSideMenu.User updateUserProfile(final long entityId, final ReqUpdateProfile reqUpdateProfile) throws RetrofitError {
         return RequestApiManager.getInstance().updateMemberProfileByProfileApi(entityId, reqUpdateProfile);
     }
 
-    public ResCommon updateMemberName(final int entityId, final ReqProfileName profileName) throws RetrofitError {
+    public ResCommon updateMemberName(final long entityId, final ReqProfileName profileName) throws RetrofitError {
         return RequestApiManager.getInstance().updateMemberNameByProfileApi(entityId, profileName);
     }
 
-    public ResLeftSideMenu.User updateMemberEmail(int entityId, String email) throws RetrofitError {
+    public ResLeftSideMenu.User updateMemberEmail(long entityId, String email) throws RetrofitError {
         return RequestApiManager.getInstance().updateMemberEmailByProfileApi(entityId, new ReqAccountEmail(email));
     }
 
@@ -182,11 +182,11 @@ public class EntityClientManager {
      * File 관련
      * **********************************************************
      */
-    public ResFileDetail getFileDetail(final int messageId) throws RetrofitError {
+    public ResFileDetail getFileDetail(final long messageId) throws RetrofitError {
         return RequestApiManager.getInstance().getFileDetailByMessagesApiAuth(selectedTeamId, messageId);
     }
 
-    public ResCommon sendMessageComment(final int messageId, String comment, List<MentionObject> mentions) throws RetrofitError {
+    public ResCommon sendMessageComment(final long messageId, String comment, List<MentionObject> mentions) throws RetrofitError {
 
         final ReqSendComment reqSendComment = new ReqSendComment(comment, mentions);
 
@@ -196,19 +196,19 @@ public class EntityClientManager {
                 reqSendComment);
     }
 
-    public ResCommon shareMessage(final int messageId, int cdpIdToBeShared) throws RetrofitError {
+    public ResCommon shareMessage(final long messageId, long cdpIdToBeShared) throws RetrofitError {
         final ReqShareMessage reqShareMessage = new ReqShareMessage();
         reqShareMessage.shareEntity = cdpIdToBeShared;
         reqShareMessage.teamId = selectedTeamId;
         return RequestApiManager.getInstance().shareMessageByMessagesApiAuth(reqShareMessage, messageId);
     }
 
-    public ResCommon unshareMessage(final int messageId, int cdpIdToBeunshared) throws RetrofitError {
+    public ResCommon unshareMessage(final long messageId, long cdpIdToBeunshared) throws RetrofitError {
         final ReqUnshareMessage reqUnshareMessage = new ReqUnshareMessage(selectedTeamId, cdpIdToBeunshared);
         return RequestApiManager.getInstance().unshareMessageByMessagesApiAuth(reqUnshareMessage, messageId);
     }
 
-//    public ResCommon modifyMessageComment(final int messageId, String comment, final int feedbackId)
+//    public ResCommon modifyMessageComment(final long messageId, String comment, final long feedbackId)
 //            throws RetrofitError {
 //        final ReqSendComment reqModifyComment = new ReqSendComment();
 //        reqModifyComment.teamId = selectedTeamId;
@@ -216,36 +216,36 @@ public class EntityClientManager {
 //        return RequestApiManager.getInstance().modifyMessageCommentByCommentsApi(reqModifyComment, feedbackId, messageId);
 //    }
 
-    public ResCommon deleteMessageComment(final int messageId, final int feedbackId) throws RetrofitError {
+    public ResCommon deleteMessageComment(final long messageId, final long feedbackId) throws RetrofitError {
         return RequestApiManager.getInstance().deleteMessageCommentByCommentsApi(selectedTeamId, feedbackId, messageId);
     }
 
-    public ResCommon deleteFile(final int fileId) throws RetrofitError {
+    public ResCommon deleteFile(final long fileId) throws RetrofitError {
         return RequestApiManager.getInstance().deleteFileByFileApi(selectedTeamId, fileId);
     }
 
-    public ResCommon modifyChannelDescription(int entityId, String description) throws RetrofitError {
+    public ResCommon modifyChannelDescription(long entityId, String description) throws RetrofitError {
         ReqModifyTopicDescription entityInfo = new ReqModifyTopicDescription();
         entityInfo.teamId = selectedTeamId;
         entityInfo.description = description;
         return RequestApiManager.getInstance().modifyPublicTopicDescriptionByChannelApi(entityInfo, entityId);
     }
 
-    public ResCommon modifyPrivateGroupDescription(int entityId, String description) throws RetrofitError {
+    public ResCommon modifyPrivateGroupDescription(long entityId, String description) throws RetrofitError {
         ReqModifyTopicDescription entityInfo = new ReqModifyTopicDescription();
         entityInfo.teamId = selectedTeamId;
         entityInfo.description = description;
         return RequestApiManager.getInstance().modifyGroupDescriptionByGroupApi(entityInfo, entityId);
     }
 
-    public ResCommon modifyChannelAutoJoin(int entityId, boolean autoJoin) {
+    public ResCommon modifyChannelAutoJoin(long entityId, boolean autoJoin) {
         ReqModifyTopicAutoJoin topicAutoJoin = new ReqModifyTopicAutoJoin();
         topicAutoJoin.teamId = selectedTeamId;
         topicAutoJoin.autoJoin = autoJoin;
         return RequestApiManager.getInstance().modifyPublicTopicAutoJoinByChannelApi(topicAutoJoin, entityId);
     }
 
-    public int getSelectedTeamId() {
+    public long getSelectedTeamId() {
         return selectedTeamId;
     }
 }
