+++ app/src/main/java/com/tosslab/jandi/app/network/client/privatetopic/IGroupApiLoader.java
@@ -15,13 +15,13 @@ public interface IGroupApiLoader {
 
     IExecutor<ResCommon> loadCreatePrivateGroupByGroupApi(ReqCreateTopic group);
 
-    IExecutor<ResCommon> loadModifyGroupByGroupApi(ReqModifyTopicName channel, int groupId);
+    IExecutor<ResCommon> loadModifyGroupByGroupApi(ReqModifyTopicName channel, long groupId);
 
-    IExecutor<ResCommon> loadModifyGroupDescriptionByGroupApi(ReqModifyTopicDescription entityInfo, int entityId);
+    IExecutor<ResCommon> loadModifyGroupDescriptionByGroupApi(ReqModifyTopicDescription entityInfo, long entityId);
 
-    IExecutor<ResCommon> loadDeleteGroupByGroupApi(int teamId, int groupId);
+    IExecutor<ResCommon> loadDeleteGroupByGroupApi(long teamId, long groupId);
 
-    IExecutor<ResCommon> loadLeaveGroupByGroupApi(int groupId, ReqTeam team);
+    IExecutor<ResCommon> loadLeaveGroupByGroupApi(long groupId, ReqTeam team);
 
-    IExecutor<ResCommon> loadInviteGroupByGroupApi(int groupId, ReqInviteTopicUsers inviteUsers);
+    IExecutor<ResCommon> loadInviteGroupByGroupApi(long groupId, ReqInviteTopicUsers inviteUsers);
 }
