+++ app/src/main/java/com/tosslab/jandi/app/network/models/ReqSetMarker.java
@@ -8,11 +8,11 @@ public class ReqSetMarker {
     public static final String PRIVATEGROUP = "privateGroup";
     public static final String USER = "user";
 
-    public int teamId;
-    public int lastLinkId;
+    public long teamId;
+    public long lastLinkId;
     public String entityType;
 
-    public ReqSetMarker(int teamId, int lastLinkId, String entityType) {
+    public ReqSetMarker(long teamId, long lastLinkId, String entityType) {
         this.teamId = teamId;
         this.lastLinkId = lastLinkId;
         this.entityType = entityType;
